requirements.txt
Flask==2.3.3
python-telegram-bot==20.6
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23